% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgInR-package.r
\docType{package}
\name{ProgInR}
\alias{ProgInR}
\alias{ProgInR-package}
\title{ProgInR}
\description{
This package provides some basic functionality to help BAHCM Reproducible
Programming in R students.
}
