% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation loop}
\usage{
run_simulation(step_function, latest.df, end.time, ...)
}
\arguments{
\item{step_function}{- function to run a timestep (step_function()) which returns
a list containing updated.pop with the updated population and
end.experiment which is TRUE if the experiment has ended, OR which just
returns a data frame with the updated population}

\item{latest.df}{- data frame with columns corresponding to function requirements}

\item{end.time}{- end time of simulation}

\item{...}{- any other arguments for step_function(), e.g. parameters or timestep}
}
\value{
output of final step of simulation, including all of the time steps up to now
}
\description{
A generic function to run a simulation loop for a fixed period of time.
}
\examples{

growth <- function(latest.df, growth.rate) {
current.count <- latest.df$count
growth.num <- current.count * growth.rate
next.count <- current.count + growth.num
next.time <- latest.df$time + 1
new.df <- data.frame(time=next.time, count=next.count)
finished <- next.count == 0
list(updated.pop=new.df, end.experiment=finished)
}
df <- data.frame(time=0, count=1)
results <- run_simulation(growth, df, 100, growth.rate=0.1)
plot_populations(results)

}

